rename to PullRequest score
